feat: convert Icon component from JS to TypeScript
- Converted from JSX to TSX with native TypeScript types
- Removed PropTypes in favor of TypeScript interfaces
- Replaced defaultProps with default parameter values
- Added proper TypeScript interfaces for IconProps and SvgAttrs
- Removed obsolete index.d.ts file
- Fixed TypeScript compilation errors
